+++ modules/default/currentweather/currentweather.js
@@ -42,6 +42,7 @@ Module.register("currentweather",{
 
 		appendLocationNameToHeader: true,
 		calendarClass: "calendar",
+		tableClass: "large",
 
 		onlyTemp: false,
 		hideTemp: false,
@@ -175,6 +176,7 @@ Module.register("currentweather",{
 	// Override dom generator.
 	getDom: function() {
 		var wrapper = document.createElement("div");
+		wrapper.className = this.config.tableClass;
 
 		if (this.config.appid === "") {
 			wrapper.innerHTML = "Please set the correct openweather <i>appid</i> in the config for module: " + this.name + ".";
@@ -193,7 +195,7 @@ Module.register("currentweather",{
 		}
 
 		var large = document.createElement("div");
-		large.className = "large light";
+		large.className = "light";
 
 		var degreeLabel = "";
 		if (this.config.units === "metric" || this.config.units === "imperial") {